Add MaestroCoreHealthMonitor
