+++ Directory.Packages.props
@@ -23,6 +23,7 @@
     <PackageVersion Include="Azure.Storage.Queues" Version="12.11.0" />
     <PackageVersion Include="Castle.Core" Version="4.3.0" />
     <PackageVersion Include="CommandLineParser" Version="2.2.1" />
+    <PackageVersion Include="CoreHealthMonitor" Version="$(CoreHealthMonitorVersion)" />
     <PackageVersion Include="EntityFrameworkCore.Triggers" Version="1.1.1" />
     <PackageVersion Include="FluentAssertions" Version="5.10.3" />
     <PackageVersion Include="FluentValidation.AspNetCore" Version="8.6.2" />
@@ -82,7 +83,7 @@
     <PackageVersion Include="Microsoft.Data.SqlClient" Version="2.1.4" />
     <PackageVersion Include="Microsoft.DncEng.CommandLineLib" Version="$(MicrosoftDncEngCommandLineLibVersion)" />
     <PackageVersion Include="Microsoft.DncEng.Configuration.Extensions" Version="$(MicrosoftDncEngConfigurationExtensionsVersion)" />
-    <PackageVersion Include="Microsoft.DotNet.Authentication.Algorithms" Version="$(MicrosoftDotNetAuthenticationAlgorithmsVersion)" /> 
+    <PackageVersion Include="Microsoft.DotNet.Authentication.Algorithms" Version="$(MicrosoftDotNetAuthenticationAlgorithmsVersion)" />
     <PackageVersion Include="Microsoft.DotNet.GitHub.Authentication" Version="$(MicrosoftDotNetGitHubAuthenticationVersion)" />
     <PackageVersion Include="Microsoft.DotNet.Internal.DependencyInjection" Version="$(MicrosoftDotNetInternalDependencyInjectionVersion)" />
     <PackageVersion Include="Microsoft.DotNet.Internal.DependencyInjection.Testing" Version="$(MicrosoftDotNetInternalDependencyInjectionTestingVersion)" />
@@ -119,8 +120,8 @@
     <PackageVersion Include="Microsoft.Extensions.Logging.ApplicationInsights" Version="2.18.0" />
     <PackageVersion Include="Microsoft.Extensions.Logging.Console" Version="6.0.0" />
     <PackageVersion Include="Microsoft.Extensions.Logging.Debug" Version="6.0.0" />
-    <PackageVersion Include="Microsoft.Extensions.Options" Version="6.0.0" /> 
-    <PackageVersion Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="6.0.0" /> 
+    <PackageVersion Include="Microsoft.Extensions.Options" Version="6.0.0" />
+    <PackageVersion Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="6.0.0" />
     <PackageVersion Include="Microsoft.Identity.Client" Version="4.48.0" />
     <PackageVersion Include="Microsoft.IdentityModel.Clients.ActiveDirectory" Version="5.2.7" />
     <PackageVersion Include="Microsoft.IdentityModel.Tokens" Version="6.17.0" />
@@ -146,7 +147,7 @@
     <PackageVersion Include="NuGet.Packaging" Version="5.11.3" />
     <PackageVersion Include="NuGet.Protocol" Version="5.11.3" />
     <PackageVersion Include="NuGet.Versioning" Version="5.11.3" />
-    <PackageVersion Include="NUnit" Version="3.13.3" />  
+    <PackageVersion Include="NUnit" Version="3.13.3" />
     <PackageVersion Include="NUnit3TestAdapter" Version="3.16.1" />
     <PackageVersion Include="Octokit" Version="0.49.0" />
     <PackageVersion Include="Quartz" Version="3.0.6" />
@@ -165,7 +166,7 @@
     <PackageVersion Include="System.Net.Http" Version="4.3.4" />
     <PackageVersion Include="System.Security.Cryptography.Xml" Version="4.7.1" />
     <PackageVersion Include="System.Text.Encodings.Web" Version="6.0.0" />
-    <PackageVersion Include="System.ValueTuple" Version="4.5.0" /> 
+    <PackageVersion Include="System.ValueTuple" Version="4.5.0" />
     <PackageVersion Include="Verify.NUnit" Version="19.6" />
     <PackageVersion Include="WindowsAzure.Storage" Version="9.3.3" />
     <PackageVersion Include="YamlDotNet" Version="9.1.4" />
