--- src/Maestro/CoreHealthMonitor.Tests/CoreHealthMonitor.Tests.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <IsPackable>false</IsPackable>
-    <RuntimeIdentifier>win7-x64</RuntimeIdentifier>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.DotNet.ServiceFabric.ServiceHost" />
-    <PackageReference Include="Microsoft.DotNet.Internal.DependencyInjection.Testing" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\CoreHealthMonitor\CoreHealthMonitor.csproj" />
-  </ItemGroup>
-
-</Project>
