--- src/Maestro/CoreHealthMonitor/CoreHealthMonitor.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <IsServiceFabricServiceProject>True</IsServiceFabricServiceProject>
-    <ServerGarbageCollection>True</ServerGarbageCollection>
-    <RuntimeIdentifier>win7-x64</RuntimeIdentifier>
-    <TargetLatestRuntimePatch>False</TargetLatestRuntimePatch>
-    <IsPackable>true</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Content Include=".config\**">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Storage.Blobs" />
-    <PackageReference Include="JetBrains.Annotations" />
-    <PackageReference Include="Microsoft.ServiceFabric.Services" />
-    <PackageReference Include="Microsoft.DotNet.ServiceFabric.ServiceHost" />
-    <PackageReference Include="Microsoft.DotNet.Internal.Health" />
-  </ItemGroup>
-
-</Project>
