+++ src/Maestro/MaestroCoreHealthMonitor/Program.cs
@@ -1,0 +1,33 @@
+using System;
+using System.Diagnostics;
+using System.Fabric;
+using System.Threading;
+using System.Threading.Tasks;
+using CoreHealthMonitor;
+using Microsoft.DotNet.ServiceFabric.ServiceHost;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.ServiceFabric.Services.Runtime;
+
+namespace MaestroCoreHealthMonitor
+{
+    internal static class Program
+    {
+        /// <summary>
+        /// This is the entry point of the service host process.
+        /// </summary>
+        private static void Main()
+        {
+            ServiceHost.Run(
+                host =>
+                {
+                    host.RegisterStatelessService<CoreHealthMonitorService>("CoreHealthMonitorType");
+                    host.ConfigureServices(Configure);
+                });
+        }
+
+        public static void Configure(IServiceCollection services)
+        {
+            CoreHealthMonitorService.Configure(services);
+        }
+    }
+}
